requirements.txt
# Core
fastapi>=0.109.1  # CVE-2024-24762 fix
uvicorn>=0.38.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.0.0  # Required by pydantic for email validation

# FastAPI Security
python-jose[cryptography]>=3.5.0  # JWT authentication
slowapi>=0.1.9                     # Rate limiting
passlib[bcrypt]>=1.7.4             # Password hashing
pyotp>=2.9.0                       # MFA/TOTP implementation
hvac>=2.1.0                        # HashiCorp Vault client

# MCP (Model Context Protocol)
mcp>=1.22.0
asyncio>=4.0.0

# AI/LLM
openai>=1.12.0
anthropic==0.34.0
langchain==0.1.16
langchain-core>=0.1.53  # CVE-2024-10940 fix
langchain-openai==0.0.8
langchain-anthropic==0.1.4
langgraph==0.0.32
langsmith>=0.1.84  # PVE-2024-72059 fix
google-generativeai

# Vector Database
chromadb>=0.4.0  # Persistent vector store
faiss-cpu  # Legacy support

# Redis (for distributed rate limiting and Celery)
redis==4.6.0  # Compatible with Celery 5.3.4
hiredis>=2.2.0  # C parser for better performance


# Computer Vision
opencv-python-headless==4.10.0.84
Pillow>=10.2.0  # Security fixes
pytesseract==0.3.10
pdf2image==1.17.0

# Data Processing - Pinned for security
pandas>=2.0.0
numpy>=1.26.0
python-dateutil>=2.8.2
duckdb>=0.9.0
polars>=1.0.0                     # Fast DataFrame library (Rust-based)
pandera>=0.18.0                   # DataFrame schema validation
sqlglot>=23.0.0                   # SQL parsing and AST validation
pybreaker>=1.0.0                  # Circuit breaker pattern
tenacity>=8.2.0                   # Retry with backoff
openpyxl>=3.1.0
scikit-learn

# Security
pyjwt
bcrypt
psutil

# Visualization - NO VERSION PINS
plotly
matplotlib
seaborn
kaleido
scipy

# Documents
python-pptx
reportlab
Jinja2

# Database/Cache
redis
sqlalchemy>=2.0.0
alembic
psycopg2-binary  # PostgreSQL adapter
neo4j>=5.0.0  # Neo4j Knowledge Graph driver
dependency-injector>=4.41.0  # Dependency injection framework

# Task Queue
celery
flower

# HTTP/Async - NO VERSION PINS
httpx
aiohttp
requests

# File Handling
python-multipart
aiofiles

# Utilities - NO VERSION PINS
python-dotenv
loguru
tenacity
tqdm
wrapt>=1.16.0  # Needed for Python 3.13 compatibility (inspect API changes)

# Validation - NO VERSION PINS
jsonschema
pyyaml

# Monitoring
prometheus-client
sentry-sdk

# Groq
groq>=0.4.0

# Async Tasks (Celery)
celery[redis]==5.3.4
flower==2.0.1  # Celery monitoring

# Distributed Tracing (OpenTelemetry)
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-otlp==1.21.0

# HTTP Client
httpx>=0.27.1  # Required by mcp>=1.22.0


# Testing
pytest==7.4.0
pytest-asyncio==0.23.4
pytest-cov

# Code Quality & Testing
black
flake8
mypy
ipython

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
bandit>=1.7.5
safety>=2.3.0

# Web Scraping - NO VERSION PINS (fixes lxml build error)
beautifulsoup4
lxml
youtube-transcript-api
lancedb
